Add NoEmbedException class
Added a new class called NoEmbedException to handle the case when the embed element is not found. This class extends the Error class and provides a custom error message. This class will be used in the reader module to throw an exception when the embed element is not present.

Deleted App/index.ts

Removed the file App/index.ts from the project. This file was not being used and had no functionality.

Fix error message in MakeFacade.ts

Updated the error message in the MakeFacade.ts file to provide more accurate information. Previously, it was throwing an error with the incorrect method name, but now it throws an error with the correct property name.

Add reader module

Added a new module called reader.ts to handle reading data from the DOM. This module exports a function that takes a name and a type as parameters and returns the corresponding data from the DOM. If the element is not found, it throws a NoEmbedException.

Update Application class

Added a new method called withProviders to the Application class. This method allows adding multiple providers to the application. It takes an array of provider classes as a parameter and adds them to the list of providers.

Update index.ts

Removed the Http module from the index.ts file as it is no longer needed. The Http module will be a Facade in @luminix/core